requirements.txt
numpy==2.4.2
opencv_python==4.11.0.86
Pillow==12.1.0
streamlit==1.45.1
torch==2.8.0
torchvision==0.23.0
